Correct table initialization assertions
I was handling initialization incorrectly.  The
assertions I had for table_info/2 were impossible
to satisfy.  A catch-call in init_store would bury
the fatal problem of a table having the wrong
type.  The pattern-match in that function was
impossible to satisfy and would guarantee a trip
to to the catch block.  A call to create_table in
that state would fail with an error.

I put in table type assertions during init_store
to yield more useful error messages when there are
schema conflicts.  On the server's console you
will see a message like the following if there is
a schema conflict:
=SUPERVISOR REPORT==== 20-Sep-2009::11:03:39 ===
     Supervisor: {local,scrumjet_sup}
     Context:    start_error
     Reason:     {{badmatch,bag},
                  [{scrumjet_task,init_store,0},
                   {scrumjet_task,init,1},
                   {gen_server,init_it,6},
                   {proc_lib,init_p_do_apply,3}]}
     Offender:   [{pid,undefined},
                  {name,scrumjet_task},
                  {mfa,{scrumjet_task,start_link,[]}},
                  {restart_type,permanent},
                  {shutdown,5000},
                  {child_type,worker}]

 - Assert actual type of table during init

 - Catch only table no_exists not catch-all exits